Avoid looking up the schedule time for the source client, just increase by the skip value
Because of caching effects by another core running the source client, it's possible that
the update could get an old timestamp and therefore triggering this worker to become busy
thinking that the client needs rescheduling.  While some cache avoiding technique could
be used, instead I just add the adjustment the source client uses as it will be in the
same range.